polish(pu): polish katago_policy, polish test_katago_bot, fix go_env reward bug
