+++ zoo/board_games/go/config/go_alphazero_eval_config.py
@@ -15,12 +15,15 @@ if __name__ == '__main__':
     # model_path = None
 
     seeds = [0]
-    num_episodes_each_seed = 1
+    num_episodes_each_seed = 3
     # If True, you can play with the agent.
     main_config.env.agent_vs_human = False
     # main_config.env.agent_vs_human = True
+    main_config.env.save_gif_replay = True
+    main_config.env.render_in_ui = True
 
     create_config.env_manager.type = 'base'
+    main_config.env.collector_env_num = 1
     main_config.env.evaluator_env_num = 1
     main_config.env.n_evaluator_episode = 1
     total_test_episodes = num_episodes_each_seed * len(seeds)
