+++ zoo/board_games/go/envs/test_go_mcts_bot.py
@@ -1,5 +1,5 @@
 from easydict import EasyDict
-from zoo.board_games.go.envs.go_env import GoEnv
+from zoo.board_games.go.envs.go_env import GoEnv, flatten_action_to_gtp_action
 from zoo.board_games.mcts_bot import MCTSBot
 
 import pytest
@@ -9,8 +9,13 @@ import numpy as np
 cfg = EasyDict(dict(
     board_size=6,
     num_simulations=80,
-    # num_simulations=20,
     # board_size=5,
+    # num_simulations=20,
+    save_gif_replay=False,
+    render_in_ui=False,
+    katago_checkpoint_path="/Users/puyuan/code/KataGo/kata1-b18c384nbt-s6582191360-d3422816034/model.ckpt",
+    ignore_pass_if_have_other_legal_actions=True,
+    save_gif_path='./',
     komi=7.5,
     prob_random_agent=0,
     prob_expert_agent=0,
@@ -27,163 +32,194 @@ cfg = EasyDict(dict(
 @pytest.mark.envtest
 class TestGoBot:
 
-    def test_go_mcts_vs_human(self):
-        # player_0  num_simulation=1000, will win
-        # player_1  num_simulation=1
-        env = GoEnv(cfg)
-        obs = env.reset()
-        state = obs['board']
-        player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
+    # def test_go_mcts_vs_human(self):
+    #     # player_0  num_simulation=1000, will win
+    #     # player_1  num_simulation=1
+    #     env = GoEnv(cfg)
+    #     obs = env.reset()
+    #     state = obs['board']
+    #     player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
+    #
+    #     player_index = 0  # player 1 first
+    #     print('#' * 15)
+    #     print(state)
+    #     while not env.get_done_reward()[0]:
+    #         if player_index == 0:
+    #             action = player_1.get_actions(state, player_index=player_index)
+    #             player_index = 1
+    #         else:
+    #             print('-' * 40)
+    #             # action = player_2.get_actions(state, player_index=player_index)
+    #             # action = env.random_action()
+    #             action = env.human_to_action()
+    #             player_index = 0
+    #
+    #         timestep = env.step(action)
+    #         env.render('human')
+    #         # time.sleep(0.1)
+    #         state = timestep.obs['board']
+    #         print('-' * 40)
+    #         print(state)
+    #     assert env.get_done_winner()[1] == 1, f'winner is {env.get_done_winner()[1]}, player 1 should win'
+    #
+    # def test_go_mcts_vs_random(self):
+    #     # player_0  num_simulation=1000, will win
+    #     # player_1  num_simulation=1
+    #     env = GoEnv(cfg)
+    #     obs = env.reset()
+    #     state = obs['board']
+    #     player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
+    #
+    #     player_index = 0  # player 1 first
+    #     print('#' * 15)
+    #     print(state)
+    #     print('#' * 15)
+    #     while not env.get_done_reward()[0]:
+    #         if player_index == 0:
+    #             action = player_1.get_actions(state, player_index=player_index)
+    #             player_index = 1
+    #         else:
+    #             print('-' * 40)
+    #             # action = player_2.get_actions(state, player_index=player_index)
+    #             action = env.random_action()
+    #             player_index = 0
+    #
+    #         timestep = env.step(action)
+    #         # env.render('human')
+    #         # time.sleep(0.1)
+    #         state = timestep.obs['board']
+    #         print('-' * 40)
+    #         print(state)
+    #     assert env.get_done_winner()[1] == 1, f'winner is {env.get_done_winner()[1]}, player 1 should win'
+    #
+    # def test_go_self_play_mode_player1_win(self):
+    #     # player_0  num_simulation=1000, will win
+    #     # player_1  num_simulation=1
+    #     env = GoEnv(cfg)
+    #     obs = env.reset()
+    #     state = obs['board']
+    #     player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
+    #     player_2 = MCTSBot(GoEnv, cfg, 'player 2', int(cfg.num_simulations/2))  # player_index = 1, player = 2
+    #
+    #     player_index = 0  # player 1 first
+    #     print('#' * 15)
+    #     print(state)
+    #     print('#' * 15)
+    #     while not env.get_done_reward()[0]:
+    #         if player_index == 0:
+    #             action = player_1.get_actions(state, player_index=player_index)
+    #             player_index = 1
+    #         else:
+    #             print('-' * 40)
+    #             action = player_2.get_actions(state, player_index=player_index)
+    #             player_index = 0
+    #         timestep = env.step(action)
+    #         state = timestep.obs['board']
+    #         print('-' * 40)
+    #         print(state)
+    #     assert env.get_done_winner()[1] == 1, f'winner is {env.get_done_winner()[1]}, player 1 should win'
+    #
+    # def test_go_self_play_mode_player2_win(self):
+    #     # player_0  num_simulation=1
+    #     # player_1  num_simulation=1000, will win
+    #     env = GoEnv(cfg)
+    #     obs = env.reset()
+    #     state = obs['board']
+    #     player_1 = MCTSBot(GoEnv, cfg, 'player 1', 1)  # player_index = 0, player = 1
+    #     player_2 = MCTSBot(GoEnv, cfg, 'player 2', cfg.num_simulations)  # player_index = 1, player = 2
+    #
+    #     player_index = 0  # player 1 first
+    #     print('#' * 15)
+    #     print(state)
+    #     print('#' * 15)
+    #     while not env.get_done_reward()[0]:
+    #         if player_index == 0:
+    #             action = player_1.get_actions(state, player_index=player_index)
+    #             player_index = 1
+    #         else:
+    #             print('-' * 40)
+    #             action = player_2.get_actions(state, player_index=player_index)
+    #             player_index = 0
+    #         timestep = env.step(action)
+    #         state = timestep.obs['board']
+    #         print('-' * 40)
+    #         print(state)
+    #     assert env.get_done_winner()[1] == 2, f'winner is {env.get_done_winner()[1]}, player 2 should win'
+    #
+    # def test_go_self_play_mode_draw(self):
+    #     # player_0  num_simulation=1000
+    #     # player_1  num_simulation=1000, will draw
+    #     cfg.num_simulations = 50
+    #
+    #     env = GoEnv(cfg)
+    #     obs = env.reset()
+    #     state = obs['board']
+    #
+    #     player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
+    #     player_2 = MCTSBot(GoEnv, cfg, 'player 2', cfg.num_simulations)  # player_index = 1, player = 2
+    #
+    #     player_index = 0  # player 1 fist
+    #     print(state)
+    #     print('-' * 40)
+    #     while not env.get_done_reward()[0]:
+    #         if player_index == 0:
+    #             action = player_1.get_actions(state, player_index=player_index)
+    #             player_index = 1
+    #         else:
+    #             print('-' * 40)
+    #             action = player_2.get_actions(state, player_index=player_index)
+    #             player_index = 0
+    #         timestep = env.step(action)
+    #
+    #         env.render('human')
+    #         # time.sleep(0.1)
+    #
+    #         state = timestep.obs['board']
+    #         print('-' * 40)
+    #         print(state)
+    #     assert env.get_done_winner()[1] == -1, f'winner is {env.get_done_winner()[1]}, two players should draw'
+    #
 
-        player_index = 0  # player 1 first
-        print('#' * 15)
-        print(state)
-        while not env.get_done_reward()[0]:
-            if player_index == 0:
-                action = player_1.get_actions(state, player_index=player_index)
-                player_index = 1
-            else:
-                print('-' * 40)
-                # action = player_2.get_actions(state, player_index=player_index)
-                # action = env.random_action()
-                action = env.human_to_action()
-                player_index = 0
-
-            timestep = env.step(action)
-            env.render('human')
-            # time.sleep(0.1)
-            state = timestep.obs['board']
-            print('-' * 40)
-            print(state)
-        assert env.get_done_winner()[1] == 1, f'winner is {env.get_done_winner()[1]}, player 1 should win'
-
-    def test_go_mcts_vs_random(self):
-        # player_0  num_simulation=1000, will win
-        # player_1  num_simulation=1
-        env = GoEnv(cfg)
-        obs = env.reset()
-        state = obs['board']
-        player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
-
-        player_index = 0  # player 1 first
-        print('#' * 15)
-        print(state)
-        print('#' * 15)
-        while not env.get_done_reward()[0]:
-            if player_index == 0:
-                action = player_1.get_actions(state, player_index=player_index)
-                player_index = 1
-            else:
-                print('-' * 40)
-                # action = player_2.get_actions(state, player_index=player_index)
-                action = env.random_action()
-                player_index = 0
-
-            timestep = env.step(action)
-            # env.render('human')
-            # time.sleep(0.1)
-            state = timestep.obs['board']
-            print('-' * 40)
-            print(state)
-        assert env.get_done_winner()[1] == 1, f'winner is {env.get_done_winner()[1]}, player 1 should win'
-
-    def test_go_self_play_mode_player1_win(self):
-        # player_0  num_simulation=1000, will win
-        # player_1  num_simulation=1
-        env = GoEnv(cfg)
-        obs = env.reset()
-        state = obs['board']
-        player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
-        player_2 = MCTSBot(GoEnv, cfg, 'player 2', int(cfg.num_simulations/2))  # player_index = 1, player = 2
-
-        player_index = 0  # player 1 first
-        print('#' * 15)
-        print(state)
-        print('#' * 15)
-        while not env.get_done_reward()[0]:
-            if player_index == 0:
-                action = player_1.get_actions(state, player_index=player_index)
-                player_index = 1
-            else:
-                print('-' * 40)
-                action = player_2.get_actions(state, player_index=player_index)
-                player_index = 0
-            timestep = env.step(action)
-            state = timestep.obs['board']
-            print('-' * 40)
-            print(state)
-        assert env.get_done_winner()[1] == 1, f'winner is {env.get_done_winner()[1]}, player 1 should win'
-
-    def test_go_self_play_mode_player2_win(self):
-        # player_0  num_simulation=1
-        # player_1  num_simulation=1000, will win
-        env = GoEnv(cfg)
-        obs = env.reset()
-        state = obs['board']
-        player_1 = MCTSBot(GoEnv, cfg, 'player 1', 1)  # player_index = 0, player = 1
-        player_2 = MCTSBot(GoEnv, cfg, 'player 2', cfg.num_simulations)  # player_index = 1, player = 2
-
-        player_index = 0  # player 1 first
-        print('#' * 15)
-        print(state)
-        print('#' * 15)
-        while not env.get_done_reward()[0]:
-            if player_index == 0:
-                action = player_1.get_actions(state, player_index=player_index)
-                player_index = 1
-            else:
-                print('-' * 40)
-                action = player_2.get_actions(state, player_index=player_index)
-                player_index = 0
-            timestep = env.step(action)
-            state = timestep.obs['board']
-            print('-' * 40)
-            print(state)
-        assert env.get_done_winner()[1] == 2, f'winner is {env.get_done_winner()[1]}, player 2 should win'
-
-    def test_go_self_play_mode_draw(self):
-        # player_0  num_simulation=1000
-        # player_1  num_simulation=1000, will draw
-        cfg.num_simulations = 50
-
-        env = GoEnv(cfg)
-        obs = env.reset()
-        state = obs['board']
-
-        player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
-        player_2 = MCTSBot(GoEnv, cfg, 'player 2', cfg.num_simulations)  # player_index = 1, player = 2
-
-        player_index = 0  # player 1 fist
-        print(state)
-        print('-' * 40)
-        while not env.get_done_reward()[0]:
-            if player_index == 0:
-                action = player_1.get_actions(state, player_index=player_index)
-                player_index = 1
-            else:
-                print('-' * 40)
-                action = player_2.get_actions(state, player_index=player_index)
-                player_index = 0
-            timestep = env.step(action)
-
-            env.render('human')
-            # time.sleep(0.1)
-
-            state = timestep.obs['board']
-            print('-' * 40)
-            print(state)
-        assert env.get_done_winner()[1] == -1, f'winner is {env.get_done_winner()[1]}, two players should draw'
 
     def test_go_self_play_mode_case_1(self):
         env = GoEnv(cfg)
         init_state = np.array([
-            [0,  0,  0, -1, -1],
-            [0,  0,  1,  1, -1],
-            [0,  1, -1,  1, -1],
-            [0,  0,  0,  1, -1],
-            [0,  0,  0,  0,  1],
+            [-1,  -1,   1,  -1,   0,   0],
+            [ 0,   0,   0,   0,  -1,   0],
+            [ 0,   0,   0,   0,   0,  -1],
+            [ 0,   0,   0,   0,   1,   0],
+            [-1,   0,   0,  -1,  -1,   1],
+            [ 0,  -1,   0,   1,   1,   0],
         ])
 
+        # TODO
+        cfg.num_simulations = 500
+
+        player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
+        player_2 = MCTSBot(GoEnv, cfg, 'player 2', cfg.num_simulations)  # player_index = 1, player = 2
+        player_index = 0  # player 1 fist
+
+        obs = env.reset(player_index, init_state)
+        state = obs['board']
+        print(state)
+        print('#' * 15)
+
+        while not env.get_done_reward()[0]:
+            if player_index == 0:
+                action = player_1.get_actions(state, player_index=player_index)
+                print(f"mcts_gtp_action: {flatten_action_to_gtp_action(action, cfg.board_size)} for case 1")
+            break
+
+    def test_go_self_play_mode_case_2(self):
+        env = GoEnv(cfg)
+        init_state = np.array([
+            [ 0,   0,   0,   0,   0,   0],
+            [-1,   0,   0,   0,  -1,   0],
+            [ 0,  -1,   0,   0,  -1,   0],
+            [-1,  -1,   1,   1,  -1,  -1],
+            [ 1,   1,   0,   0,   1,   1],
+            [ 0,   0,   0,   0,   0,   0],
+        ])
         # TODO
         cfg.num_simulations = 50
 
@@ -199,24 +235,25 @@ class TestGoBot:
         while not env.get_done_reward()[0]:
             if player_index == 0:
                 action = player_1.get_actions(state, player_index=player_index)
-                assert action == 2
+                print(f"mcts_gtp_action: {flatten_action_to_gtp_action(action, cfg.board_size)} for case 2")
+            break
 
-    def test_go_self_play_mode_case_2(self):
+    def test_go_self_play_mode_case_3(self):
         env = GoEnv(cfg)
         init_state = np.array([
-            [0,  0,  1,  1,  1],
-            [0,  0, -1, -1,  1],
-            [0,  0,  0, -1,  1],
-            [0,  0,  0, -1,  1],
-            [0,  0,  0,  0, -1],
+            [ 1,   1,   1,   1,   0,   0],
+            [ 1,  -1,  -1,   1,   1,   1],
+            [ 1,   0,   0,  -1,  -1,   1],
+            [ 1,  -1,   0,   1,  -1,   1],
+            [ 1,  -1,   1,   1,  -1,   1],
+            [-1,  -1,  -1,  -1,  -1,   1],
         ])
-
         # TODO
-        cfg.num_simulations = 50
+        cfg.num_simulations = 500
 
         player_1 = MCTSBot(GoEnv, cfg, 'player 1', cfg.num_simulations)  # player_index = 0, player = 1
         player_2 = MCTSBot(GoEnv, cfg, 'player 2', cfg.num_simulations)  # player_index = 1, player = 2
-        player_index = 1  # player 2 fist
+        player_index = 0  # player 1 fist
 
         obs = env.reset(player_index, init_state)
         state = obs['board']
@@ -224,13 +261,13 @@ class TestGoBot:
         print('#' * 15)
 
         while not env.get_done_reward()[0]:
-            if player_index == 1:
-                action = player_2.get_actions(state, player_index=player_index)
-                assert action == 1
-
+            if player_index == 0:
+                action = player_1.get_actions(state, player_index=player_index)
+                print(f"mcts_gtp_action: {flatten_action_to_gtp_action(action, cfg.board_size)} for case 3")
+            break
 
 # test = TestGoBot().test_go_self_play_mode_player1_win()
 # test = TestGoBot().test_go_self_play_mode_draw()
-# test = TestGoBot().test_go_self_play_mode_case_2()
-test = TestGoBot().test_go_mcts_vs_human()
+# test = TestGoBot().test_go_mcts_vs_human()
 
+test = TestGoBot().test_go_self_play_mode_case_1()
