+++ zoo/board_games/go/envs/test_katago_bot.py
@@ -4,14 +4,21 @@ import numpy as np
 import pytest
 from easydict import EasyDict
 
-from zoo.board_games.go.envs.go_env import GoEnv
-from zoo.board_games.go.envs.katago_play_for_lightzero import str_coord
-
+from zoo.board_games.go.envs.go_env import GoEnv, flatten_action_to_gtp_action
+from zoo.board_games.go.envs.katago_policy_for_lightzero import str_coord
 cfg = EasyDict(
+    board_size=19,
     # board_size=9,
-    board_size=6,
-
+    # board_size=6,
+    save_gif_replay=True,
+    render_in_ui=True,
+    # save_gif_replay=False,
+    # render_in_ui=False,
+    katago_checkpoint_path="/Users/puyuan/code/KataGo/kata1-b18c384nbt-s6582191360-d3422816034/model.ckpt",
+    ignore_pass_if_have_other_legal_actions=True,
+    save_gif_path='./',
     komi=7.5,
+    # komi=7,
     battle_mode='self_play_mode',
     prob_random_agent=0,
     channel_last=False,
@@ -31,13 +38,13 @@ class TestKataGoBot:
 
         env = GoEnv(cfg)
         test_episodes = 1
-        for _ in range(test_episodes):
-            print('NOTE：actions are counted by column, such as action 9, which is the second column and the first row')
+        for i in range(test_episodes):
+            print(f'episode {i}')
+            print('='*20)
+
             obs = env.reset()
-            print(obs['observation'].shape, obs['action_mask'].shape)
-            print(obs['observation'], obs['action_mask'])
-            actions_black = [0, 2, 0]
-            actions_white = [1, 6]
+            # print(obs['observation'].shape, obs['action_mask'].shape)
+            # print(obs['observation'], obs['action_mask'])
             # env.render()
             # TODO(pu): katago_game_state init
             turn = 0
@@ -45,7 +52,11 @@ class TestKataGoBot:
                 turn += 1
                 print('turn: ', turn)
                 # ****** player 1's turn ******
+                # bot_action = env.random_action()
+                # bot_action = env.human_to_action()
+                # bot_action = env.human_to_gtp_action()
                 bot_action = env.get_katago_action(to_play=1)
+
                 if bot_action not in env.legal_actions:
                     logging.warning(
                         f"You input illegal *bot* action: {bot_action}, the legal_actions are {env.legal_actions}. "
@@ -57,9 +68,8 @@ class TestKataGoBot:
                 katago_flatten_action = env.lz_flatten_to_katago_flatten(bot_action, env.board_size)
                 print('player 1:', str_coord(katago_flatten_action, env.katago_game_state.board))
                 env.update_katago_internal_game_state(katago_flatten_action, to_play=1)
+
                 action = bot_action
-                # action = env.human_to_action()
-                # action = env.random_action()
                 # action = actions_black[i]
                 # print('player 1 (black): ', action)
                 obs, reward, done, info = env.step(action)
@@ -69,7 +79,6 @@ class TestKataGoBot:
                 assert isinstance(done, bool)
                 assert isinstance(reward, float) or isinstance(reward, int)
                 # env.render('board')
-                env.render('human')
 
                 if done:
                     if reward > 0:
@@ -81,6 +90,7 @@ class TestKataGoBot:
                     break
 
                 # ****** player 2's turn ******
+                # bot_action = env.human_to_gtp_action()
                 bot_action = env.get_katago_action(to_play=2)
                 if bot_action not in env.legal_actions:
                     logging.warning(
@@ -103,7 +113,7 @@ class TestKataGoBot:
                 # print(self.board)
                 # print(obs, reward, done, info)
                 # env.render('board')
-                env.render('human')
+
                 if done:
                     if reward > 0:
                         print('player 2 (white) win')
@@ -113,4 +123,4 @@ class TestKataGoBot:
                         print('draw')
                     break
 
-TestKataGoBot().test_katago_bot()
+# TestKataGoBot().test_katago_bot()
